Trader/projects/requirements.txt
APScheduler==3.6.3
backtrader==1.9.76.123
certifi==2020.12.5    
chardet==4.0.0        
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.2.2
numpy==1.20.2
pandas==1.2.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4